refactored, split into mutliple files
